Add new remotes backend to bravetools.
Remotes are stored in a new "remotes" folder in bravetools home dir.
A remote is a human-editable json file and associates a remote with a name, address, and protocol.
TLS auth for a remote is loaded if necessary.

For now, the only remote is the default bravetools local remote created
during init - if all works fine then next step is to expose the ability to select
different remotes to users.